requirements.txt
python-dotenv
pyTelegramBotAPI
g4f